BLGMIGRATION-922 refactor: move package
