--- core/src/main/scala/com/nulabinc/backlog/migration/common/persistence/sqlite/ops/AllTableOps.scala
@@ -1,12 +1,0 @@
-package com.nulabinc.backlog.migration.common.persistence.sqlite.ops
-
-// import monix.execution.Scheduler
-
-// case class AllTableOps()(implicit exc: Scheduler) {
-//   val backlogStatusTableOps = StatusTableOps()
-
-//   val createDatabaseOps =
-//     DBIO.seq(
-//       backlogStatusTableOps.createTable
-//     )
-// }
