--- core/src/main/scala/com/nulabinc/backlog/migration/common/persistence/sqlite/ops/StatusTableOps.scala
@@ -1,13 +1,0 @@
-package com.nulabinc.backlog.migration.common.persistence.sqlite.ops
-
-// import com.nulabinc.backlog.migration.common.persistence.sqlite.tables.{
-//   BacklogStatusRow,
-//   BacklogStatusTable
-// }
-// import monix.execution.Scheduler
-// import slick.lifted.TableQuery
-
-// case class StatusTableOps()(implicit ec: Scheduler)
-//     extends BaseTableOps[BacklogStatusRow, BacklogStatusTable] {
-//   protected val tableQuery = TableQuery[BacklogStatusTable]
-// }
